backend/requirements.txt
asgiref==3.11.1
Django==6.0.4
django-cors-headers==4.0.0
django-filter==25.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
gunicorn==23.0.0
psycopg2-binary==2.9.11
python-dateutil==2.9.0.post0
pytz==2026.1.post1
sqlparse==0.5.5
tzdata==2026.1
whitenoise==6.11.0
dj-database-url==3.0.1
